+++ app/jobs/intent_reminder_job.rb
@@ -10,12 +10,13 @@ class IntentReminderJob < ApplicationJob
   def perform(draw:)
     students = draw.students.joins(:draw_membership)
                    .where(draw_memberships: { intent: %w(undeclared) })
-    students.each { |s| send_email(s) }
+    admins = College.current.users.admin
+    (students + admins).each { |s| send_email(user: s, draw: draw) }
   end
 
   private
 
-  def send_email(user)
-    StudentMailer.intent_reminder(user: user).deliver_now
+  def send_email(user:, draw:)
+    StudentMailer.intent_reminder(user: user, draw: draw).deliver_now
   end
 end