+++ app/services/draw_activator.rb
@@ -57,8 +57,9 @@ class DrawActivator
   end
 
   def send_emails
-    draw.students.each do |student|
-      mailer.draw_invitation(user: student, college: college).deliver_later
+    (draw.students + college.users.admin).each do |student|
+      mailer.draw_invitation(user: student, draw: draw, college: college)
+            .deliver_later
     end
   end