+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## v2.2.0 - 2019-02-12
 ### Added
 * Add the ability to assign lottery numbers by size ([#848](https://gitlab.com/yale-sdmp/vesta/issues/848)).
+* Add admins to all bulk user emails ([#887](https://gitlab.com/yale-sdmp/vesta/issues/887)).
 
 ### Fixed
 * Scope the results view and export to only return active students in the current college ([#885](https://gitlab.com/yale-sdmp/vesta/issues/885)).